# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/purchase_order_items:
    get:
      tags:
        - PurchaseOrderItems
      summary: List PurchaseOrderItems
      description: Get a paginated list of PurchaseOrderItems with optional filtering and sorting
      operationId: listPurchaseOrderItems
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseOrderItemList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PurchaseOrderItems
      summary: Create PurchaseOrderItem
      description: Create a new PurchaseOrderItem
      operationId: createPurchaseOrderItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePurchaseOrderItemRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseOrderItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/request_for_quotations:
    get:
      tags:
        - RequestForQuotations
      summary: List RequestForQuotations
      description: Get a paginated list of RequestForQuotations with optional filtering and sorting
      operationId: listRequestForQuotations
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestForQuotationList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - RequestForQuotations
      summary: Create RequestForQuotation
      description: Create a new RequestForQuotation
      operationId: createRequestForQuotation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequestForQuotationRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestForQuotation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    PurchaseOrderItem:
      type: object
      required:
        - id
        - order_id
        - company_id
        - item_id
        - quantity
        - rate
        - line_amount
        - received_qty
        - billed_qty
        - qty_received_method
        - purchase_method
        - metadata
      properties:
        id:
          type: string
          format: uuid
        order_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        warehouse_id:
          nullable: true
          type: string
          format: uuid
        description:
          nullable: true
          type: string
          maxLength: 300
        quantity:
          type: string
        rate:
          type: string
        line_amount:
          type: string
        received_qty:
          type: string
        billed_qty:
          type: string
        qty_received_method:
          type: string
        purchase_method:
          type: string
        metadata:
          type: object
...
    CreatePurchaseOrderItemRequest:
      type: object
      required:
        - order_id
        - company_id
        - item_id
        - quantity
        - rate
        - line_amount
        - received_qty
        - billed_qty
        - qty_received_method
        - purchase_method
        - metadata
      properties:
        order_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        warehouse_id:
          nullable: true
          type: string
          format: uuid
        description:
          nullable: true
          type: string
          maxLength: 300
        quantity:
          type: string
        rate:
          type: string
        line_amount:
          type: string
        received_qty:
          type: string
        billed_qty:
          type: string
        qty_received_method:
          type: string
        purchase_method:
          type: string
        metadata:
          type: object
...
    PurchaseOrderItemList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PurchaseOrderItem'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    RequestForQuotation:
      type: object
      required:
        - id
        - rfq_number
        - company_id
        - status
        - rfq_date
        - metadata
      properties:
        id:
          type: string
          format: uuid
        rfq_number:
          type: string
          maxLength: 40
        material_request_id:
          nullable: true
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        status:
          type: string
        rfq_date:
          type: string
          format: date
        response_due:
          nullable: true
          type: string
          format: date
        metadata:
          type: object
...
    CreateRequestForQuotationRequest:
      type: object
      required:
        - rfq_number
        - company_id
        - status
        - rfq_date
        - metadata
      properties:
        rfq_number:
          type: string
          maxLength: 40
        material_request_id:
          nullable: true
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        status:
          type: string
        rfq_date:
          type: string
          format: date
        response_due:
          nullable: true
          type: string
          format: date
        metadata:
          type: object
...
    RequestForQuotationList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RequestForQuotation'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []